requirements/requirements-tests.txt
black==19.3b0
flake8==3.7.8
isort==4.3.3
pytest-cov==2.7.1
